Initial commit and add commentary to readme file
In the initial commit, I only added some commentary to the README.md
file, explaining the premise of the project and establishing my
expectations for how I think the project will go. When the project is
completed, I will update the README.md file again with my closing
thoughts and review.